+++ src/lib/components/ui/button-radio-group/index.ts
@@ -1,0 +1,9 @@
+import { RadioGroup } from 'bits-ui';
+import ButtonRadioGroupItem from './button-radio-group-item.svelte';
+
+const Root = RadioGroup.Root;
+
+export {
+  Root,
+  ButtonRadioGroupItem as Item,
+};